+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 ### 0.0.2 (2022-07-09)
 * Numbers arrangement
 * Bomb alert message
+
+### 0.0.3 (2022-11-05)
+* Work with the blank spaces
+* Use and array system for the numbers and bombs arrangement